Reduce user menu hover target
Before this commit, you could hover over the invisible parts of the drop
down menu to make it visible. That was a problem because the invisible
part would cover the "hot/new/top" links on small screens, making them
unclickable. Now, only the username/avatar can be hovered over to
trigger the menu.