Redesign create-dweet UI to be hidden by default
This commit hides the UI for creating a new dweet by default, so the
first you see when entering the page is now instead the top dweet. To
show the UI, you must click the 'create a new dweet' button at the top
of the feed.

The UI for creating dweet is also cleaned up a little, to look more
similar to dweet editing and remove visual noise.

This fixes #3.