Redesign and cleanup header
This reduces the height of the header by 50px, making the content of the
page more prominent. It also sets the max-width of the header to the
industry-standard 960px to avoid elements being pushed into the extreme
left and right of large browser windows.

Finally, it reduces the font size of the dropdown menu so every item in it
can fit on one line.