Refactor theme system to reduce repetition
This should also fix the issue with switching themes in compressed mode,
since the whole link tag for themes is now loaded client-side. Also this
removes the flash of unstyled content that at least I see on my machine
when loading the site and having the dark theme enabled (in the future
this may be improved further by switching from local storage to cookies
and doing the theming server side). Anyway, with this change you will
see the light theme by default and then switch to the dark theme as soon
as it is loaded, instead of seeing nothing while waiting.

Also, this makes it much easier to make changes to the css in the
future, because not all of the styles have to be updated twice for every
change.

Finally this fixes some text that was invisible in the dark theme.

This fixes #294.