A couple of low-hanging mobile layout fixes
This reduces some padding and also tidies up the spacing in the header
by aligning the title and menu to the same grid as the rest of the site.

It should ensure that the header fits on two lines on device widths down
to 320pt (the width of an iPhone SE) on both the front page and at least
profile pages with medium-long usernames.

Also, the links under a dweet are now allowed to wrap over multiple
lines, so they don't have to all try and squish into each other in the
same line anymore.

This fixes #303.