Repair positioning of dropdown menu
With changes to the dark theme, the margins of the dropdown were
updated, but I couldn't figure out why that change was made. Anyway,
this restores the hidden dropdown menu to waaaay off screen so it can
not be clicked when hidden. I also halved the speed of the animation and
adjusted the positioning so it perfectly aligns with the username and
avatar.

This should resolve issues on both desktop on mobile where you try and
press a link somewhere on the page and accidentally click something in
the ghost menu instead.