Merge sharing options into popover again
This creates a new, slightly nicer looking, popover that contains both
the share URL and an embed code. The plan is to re-use this popover
style for more types of content on the dweet card.

When I previously did this in f73d5b0 I was not aware that the code used
for lazy-loaded dweets was completely separate in another js file.
So this time, I delete the duplicated code in scrolling.js and instead
implement it one way by binding event listeners once on body instead of
separately per dweet card. This is also better for performance when
having lots of dweets loaded.