Prevent XSS in the comment section
This patch fixes the XSS vulnerability in the comment section.

Because insert_code_blocks() marks the input text as `safe`,it is
required to escape the text manually before passing to the filter.
Right now, we don't do so.

This results in allowing users to execute arbitrary JavaScript code
through malformed comments.