Use css transforms for header animation
This improves the performance of the animation, because animating the
transform property is much faster than animating the top propoerty. The
top property triggers expensive layout calculations every time it
changes, while the transform property has been specifically optimized
for fast animations. This should ensure that the header on mobile
scrolls smoothly in and out of view.

https://www.html5rocks.com/en/tutorials/speed/high-performance-animations/