Add Math.floor() to the frame timing check
I was getting frame timeouts when the time was off by 0.3ms etc. Looks
like in some browsers the timing is given with a resolution of 1ms, so
comparing to a value with sub-ms resolution gave some judder.

This change also uses the timing passed in by requestAnimationFrame
rather than Date.now. This should be more stable as this is a time
that is sampled just before the browser starts doing all the redraws,
so it doesn't depend on how much compute has been done before the
dweet callback is called.