Fully hardcode partially hardcoded urls after weird test fail
For some reasons reverse() is failing in the @signal on Jenkins when
testing, but not while running tests locally. Seems like the signal is
called before urls.py is loaded somehow, but I don't understand why that
would be different locally.

As the urls were partially hardcoded anyway, reverse() isn't providing a
lot of value, and this text is likely to get changed later, I'm just
hardcoding the full url