Make remixing slightly more user-friendly
By showing the remix button on focus rather than on code change it
should be easier to find.

Current behavior:
The button to post a remix only shows up when you have changed the
dweet.

New behavior:
When a user clicks the code (ie. focus change) the remix button shows
up, but it starts out disabled.
The field you enter a comment in tells the user that he should edit the
code.
Once the code has changed, the remix button is enabled and the
text in the comment field changes to "Add a caption here..."