Split our dweet reply info to separate query
Rather than joining on the dweet and user table one extra time in the
main query (which needs to be sorted and is already quite huge), this
splits out the query for loading information about the dweet to a
separate query. That means the total number of SQL queries on the
frontpage will increase by one, but I believe the total execution time
and amount of data to load from the database will be smaller, since the
data is joined in python instead.