Only annotate num_likes on Top views
The num_likes annotation is quite heavy, so avoiding it on pages where
it is not used seems smart.

With this, only the TopAllDweetFeed view should be particularly slow,
hopefully (the rest of the top views are fast since they start by
filtering out a lot of data based on posted at date before doing their
sorting).