Dweet Perf - Define "u" with "new Function()" constructor
Defining the `u` function with eval seriously degrades JS performance,
particular in chrome's V8 which seems to be completely unoptimised.
There are alternative ways to continue using eval which provide better
performance, however none are ideal and results are inconsistent between
JS engines. Therefore we should use the function constructor instead
which is reliably optimised.

As per the previous comments removed by this commit, the function
constructor was previously avoided in order to allow dweet code to
safely overwrite the global `u` variable without breaking the loop's
reference to `u`. We can accommodate this by additionally assigning `u`
to an undesirably long variable name `userFunction` which will not be
accidentally overwritten.