Dweet Perf - Decompress unicode packed dweets
The previous commit only fixes performance issues related to using eval
for the function definition. Although less severe, using eval inside of
the function body also has a negative impact on performance, and unicode
packing commonly used to achieve 185 or 194 byte dweets use eval.

We can safely identify these and replace them with the unpacked code for
much better performance. There are a few variants, but they all share a
common signature that is simple enough to be identified with a regex.
If a variant is not matched it will not be touched.