Dweet Perf - Reduce loopbuster check frequency
Calls to `Date.now()` or `performance.now()` can be expensive in some
browsers such as Firefox when called in a tight loop. This particularly
affects raymarchers and fractals where the inner loop will substantially
magnify the number of `Date.now()` calls.

We can reduce the frequency of calls to Date.now by counting iteration
throughout the frame and checking every n iterations. This does not
noticeably affect interrupt delay since long running loops tend to be
caused by sheer number of iterations.

Since it doesn't matter which loop we are counting and it's better to
track the time from the start of the frame rather than a loop, we can do
away with external function calls and inline all of the timer checks and
counter checks using a global timer and counter.