Dweet Perf Fix - Function compatibility
Some dweets exploited the old eval function definition string by
appending a couple }{ to create an initialisation block. This is broken
with the new function constructor method, and although it's not
officially part of the dwitter interface it would be nice to retain
compatibility with these dweets.

KilledByAPixel suggested we fall back to using the old method when the
function definition fails before finally throwing the function
definition in a nested try catch. This should nicely catch any other
possible quirks dweets were exploiting in the old eval string.