Dweet Perf Fix - Variable compatibility
The old eval method exposed the function body of `u` to the local scope
of the containing function, this is not the case for the new function
constructor which only exposes the body to the global scope.

Some dweets relied on the local `code` variable (various quines), so we
should add this to the window object to re-expose it.