Remove awesome_count from remix serializer
This serialized view of the dweet is only used to display the "remix of
{dweet}" view, and doesn't need to show how many awesomes said dweet
has.

Motivation for removing this was simply that as written, this caused one
additional query for every dweet in a dweet-list. Since this data isn't
used I simply removed it rather than finding a way to prefetch it.